+++ src/autocomplete_test.js
@@ -33,6 +33,7 @@ function initEditor(value) {
     }
     editor = ace.edit(null, {
         value: value,
+        minLines: 10,
         maxLines: 10,
         enableBasicAutocompletion: true,
         enableLiveAutocompletion: true
@@ -66,16 +67,16 @@ module.exports = {
         assert.ok(!editor.container.querySelector("style"));
 
         sendKey("a");
-        checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="2" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_completion-highlight">a</s><s "ace_">rraysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d><d "ace_line"><s "ace_completion-highlight">a</s><s "ace_">looooooooooooooooooooooooooooong_word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+        checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-roledescription="item" aria-label="arraysort" aria-setsize="2" aria-posinset="1" aria-describedby="doc-tooltip" aria-selected="true"><s "ace_completion-highlight">a</s><s "ace_">rraysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d><d "ace_line"><s "ace_completion-highlight">a</s><s "ace_">looooooooooooooooooooooooooooong_word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
             sendKey("rr");
-            checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_completion-highlight">arr</s><s "ace_">aysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+            checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-roledescription="item" aria-label="arraysort" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip" aria-selected="true"><s "ace_completion-highlight">arr</s><s "ace_">aysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
                 sendKey("r");
-                checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_completion-highlight">arr</s><s "ace_">ayso</s><s "ace_completion-highlight">r</s><s "ace_">t</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+                checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-roledescription="item" aria-label="arraysort" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip" aria-selected="true"><s "ace_completion-highlight">arr</s><s "ace_">ayso</s><s "ace_completion-highlight">r</s><s "ace_">t</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
                     
                     sendKey("Return");
                     assert.equal(editor.getValue(), "arraysort\narraysort alooooooooooooooooooooooooooooong_word");
                     editor.execCommand("insertstring", " looooooooooooooooooooooooooooong_");
-                    checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="alooooooooooooooooooooooooooooong_word" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_">a</s><s "ace_completion-highlight">looooooooooooooooooooooooooooong_</s><s "ace_">word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+                    checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-roledescription="item" aria-label="alooooooooooooooooooooooooooooong_word" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip" aria-selected="true"><s "ace_">a</s><s "ace_completion-highlight">looooooooooooooooooooooooooooong_</s><s "ace_">word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
                         sendKey("Return");
                         editor.destroy();
                         editor.container.remove();
@@ -857,16 +858,15 @@ module.exports = {
         var inline = completer.inlineRenderer;
 
         // Popup should be open, with inline text renderered.
-        assert.equal(editor.completer.popup.isOpen, true);  
+        assert.equal(completer.popup.isOpen, true);  
         assert.equal(completer.popup.getRow(), 0);
         assert.strictEqual(inline.isOpen(), true);
         assert.strictEqual(editor.renderer.$ghostText.text, "function\nthat does something\ncool");
 
-        editor.completer.popup.renderer.$loop._flush();
-        var popupTextLayer = completer.popup.renderer.$textLayer;
+        completer.popup.renderer.$loop._flush();
 
         // aria-describedby of selected popup item should have aria-describedby set to the offscreen inline screen reader div and doc-tooltip.
-        assert.strictEqual(popupTextLayer.selectedNode.getAttribute("aria-describedby"), "doc-tooltip ace-inline-screenreader-line-0 ace-inline-screenreader-line-1 ace-inline-screenreader-line-2 ");
+        assert.strictEqual(completer.popup.selectedNode.getAttribute("aria-describedby"), "doc-tooltip ace-inline-screenreader-line-0 ace-inline-screenreader-line-1 ace-inline-screenreader-line-2 ");
 
         // The elements with these IDs should have the correct content.
         assert.strictEqual(document.getElementById("ace-inline-screenreader-line-0").textContent,"function");
@@ -989,6 +989,7 @@ module.exports = {
         
         var completer = Autocomplete.for(editor);
         completer.stickySelectionDelay = 100;
+        completer.showLoadingState = true;
         user.type("Ctrl-Space");
         assert.ok(!(completer.popup && completer.popup.isOpen));
 
@@ -1005,8 +1006,11 @@ module.exports = {
 
                 editor.completers = [fastCompleter, slowCompleter];
                 user.type("Ctrl-Space");
-                assert.equal(completer.popup.data.length, 3); 
+                assert.equal(completer.popup.data.length, 4); 
+
+                // Should have top row saying 'Loading...' together with results.
                 assert.ok(isLoading());
+                assert.equal(completer.popup.data[0].caption, "Loading..."); 
                 setTimeout(() => {
                     completer.popup.renderer.$loop._flush();
                     assert.equal(completer.popup.data.length, 5);
@@ -1019,6 +1023,323 @@ module.exports = {
         function isLoading() {
             return completer.popup.renderer.container.classList.contains("ace_loading");
         }
+    },
+    "test: should not display loading state on no suggestion state": function(done) {
+        var editor = initEditor("hello world\n");
+        
+        var slowCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "slow option 1",
+                        value: "s1",
+                        score: 3
+                    }, {
+                        caption: "slow option 2",
+                        value: "s2",
+                        score: 0
+                    }
+                ];
+                setTimeout(() => {
+                    callback(null,  completions);
+                }, 200);
+            }
+        };
+
+        editor.completers = [slowCompleter];
+        
+        var completer = Autocomplete.for(editor);
+        completer.stickySelectionDelay = 100;
+        completer.emptyMessage = "no completions";
+        completer.showLoadingState = true;
+
+        user.type("doesntmatchanything");
+        user.type("Ctrl-Space");
+        assert.ok(!(completer.popup && completer.popup.isOpen));
+
+        setTimeout(() => {
+            completer.popup.renderer.$loop._flush();
+            assert.equal(completer.popup.data.length, 1);
+            assert.ok(isLoading());
+            setTimeout(() => {
+                // Should show no suggestions state without loading indicator
+                assert.equal(completer.popup.data.length, 1); 
+                assert.equal(completer.popup.data[0].caption, "no completions");
+                assert.ok(!isLoading());
+    
+                done();
+            }, 150);
+        }, 100);
+        
+        function isLoading() {
+            return completer.popup.renderer.container.classList.contains("ace_loading");
+        }
+    },
+    "test: should display ghost text after loading state if inline preview enabled": function(done) {
+        var editor = initEditor("hello world\n");
+        
+        var slowCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "slow option 1",
+                        value: "s1",
+                        score: 3
+                    }, {
+                        caption: "slow option 2",
+                        value: "s2",
+                        score: 0
+                    }
+                ];
+                setTimeout(() => {
+                    callback(null,  completions);
+                }, 200);
+            }
+        };
+
+        editor.completers = [slowCompleter];
+        
+        var completer = Autocomplete.for(editor);
+        completer.stickySelectionDelay = 100;
+        completer.inlineEnabled = true;
+        completer.showLoadingState = true;
+
+        user.type("Ctrl-Space");
+        assert.ok(!(completer.popup && completer.popup.isOpen));
+
+        setTimeout(() => {
+            completer.popup.renderer.$loop._flush();
+            assert.equal(completer.popup.data.length, 1);
+            assert.ok(isLoading());
+            setTimeout(() => {
+                assert.equal(completer.popup.data.length, 2); 
+                assert.ok(!isLoading());
+                
+                assert.strictEqual(editor.renderer.$ghostText.text, "s1");
+                done();
+            }, 150);
+        }, 100);
+        
+        function isLoading() {
+            return completer.popup.renderer.container.classList.contains("ace_loading");
+        }
+    },
+    "test: when completion gets inserted and call the onInsert method": function (done) {
+        var isInserted = false;
+        var editor = initEditor("hello world");
+        var completer = {
+            onInsert: function (_editor, el) {
+                assert.ok(!isInserted, "should not have inserted something already");
+                isInserted = el.value === "one";
+            },
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "option 1",
+                        value: "one",
+                        completer
+                    }, {
+                        caption: "option 2",
+                        value: "two",
+                        completer
+                    }, {
+                        caption: "option 3",
+                        value: "three",
+                        completer
+                    }
+                ];
+                callback(null, completions);
+            }
+        };
+        editor.completers = [completer];
+        user.type("Ctrl-Space");
+        editor.completer.popup.renderer.$loop._flush();
+        assert.equal(editor.completer.popup.isOpen, true);     
+        assert.equal(editor.completer.popup.getRow(), 0);
+        user.type("Return");
+
+        assert.ok(isInserted);
+
+        done();
+    },
+    "test: when completions get shown, call the onSeen method": function (done) {
+        var seen = [false, false, false];
+        var editor = initEditor("hello world");
+        var completer = {
+            onSeen: function (_editor, el) {
+                const index = ["one", "two", "three"].indexOf(el.value);
+                if (index >= 0) {
+                    assert.ok(!seen[index], "should not be called double");
+                    seen[index] = true;
+                }
+            },
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "option 1",
+                        value: "one",
+                        completer
+                    }, {
+                        caption: "option 2",
+                        value: "two",
+                        completer
+                    }, {
+                        caption: "option 3",
+                        value: "three",
+                        completer
+                    }
+                ];
+                callback(null, completions);
+            }
+        };
+        editor.completers = [
+            completer
+        ];
+        Autocomplete.for(editor).inlineEnabled = true;
+        user.type("Ctrl-Space");
+        editor.completer.popup.renderer.$loop._flush();
+        assert.equal(editor.completer.popup.isOpen, true);     
+        assert.equal(editor.completer.popup.getRow(), 0);
+        assert.deepEqual(seen, [true, false, false]);
+        done();
+    },
+    "test: when inline completions get shown, call the onSeen method": function (done) {
+        var seen = [false, false, false];
+        var calledDouble = false;
+        var editor = initEditor("hello world");
+        var completer = {
+            onSeen: function (_editor, el) {
+                const index = ["one", "two", "three"].indexOf(el.value);
+                if (index >= 0) {
+                    if (seen[index]) calledDouble = true;
+                    seen[index] = true;
+                }
+            },
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "option 1",
+                        value: "one",
+                        completer
+                    }, {
+                        caption: "option 2",
+                        value: "two",
+                        completer
+                    }, {
+                        caption: "option 3",
+                        value: "three",
+                        completer
+                    }
+                ];
+                callback(null, completions);
+            }
+        };
+        editor.completers = [
+            completer
+        ];
+        user.type("Ctrl-Space");
+        editor.completer.popup.renderer.$loop._flush();
+        assert.equal(editor.completer.popup.isOpen, true);     
+        assert.equal(editor.completer.popup.getRow(), 0);
+        assert.deepEqual(seen, [true, true, true]);
+        assert.ok(!calledDouble);
+        done();
+    },
+    "test: if there is very long ghost text, popup should be rendered at the bottom of the editor container": function(done) {
+        var editor = initEditor("hello world\n");
+
+        // Give enough space for the popup to appear below the editor
+        var initialDocumentHeight = document.body.style.height;
+        document.body.style.height = editor.container.getBoundingClientRect().height + 200 + "px";
+        editor.renderer.$loop._flush();
+
+        var longCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "loooooong",
+                        value: "line\n".repeat(20),
+                        score: 0
+                    }
+                ];
+                callback(null,  completions);
+            }
+        };
+
+        editor.completers = [longCompleter];
+        
+        var completer = Autocomplete.for(editor);
+        completer.inlineEnabled = true;
+
+        user.type("Ctrl-Space");
+        assert.ok(completer.popup && completer.popup.isOpen);
+
+        // Wait to account for the renderer scrolling for the virtual renderer
+        setTimeout(() => {
+            completer.popup.renderer.$loop._flush();
+
+            // Popup should start one pixel below the bottom of the editor container
+            assert.equal(
+                completer.popup.container.getBoundingClientRect().top,
+                editor.container.getBoundingClientRect().bottom + 1
+            );
+
+            // Reset back to initial values
+            document.body.style.height = initialDocumentHeight;
+            editor.renderer.$loop._flush();
+
+            done();
+        }, 100);
+    },
+    "test: if there is ghost text, popup should be rendered at the bottom of the ghost text": function(done) {
+        var editor = initEditor("");
+
+        var longCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "loooooong",
+                        value: "line\n".repeat(5),
+                        score: 0
+                    }
+                ];
+                callback(null,  completions);
+            }
+        };
+
+        editor.completers = [longCompleter];
+        
+        var completer = Autocomplete.for(editor);
+        completer.inlineEnabled = true;
+
+        user.type("Ctrl-Space");
+        assert.ok(completer.popup && completer.popup.isOpen);
+
+        // Wait to account for the renderer scrolling for the virtual renderer
+        setTimeout(() => {
+            completer.popup.renderer.$loop._flush();
+
+            // Popup should start one pixel below the bottom of the ghost text
+            assert.equal(
+                completer.popup.container.getBoundingClientRect().top,
+                editor.renderer.$ghostTextWidget.el.getBoundingClientRect().bottom + 1
+            );
+
+            done();
+        }, 100);
+    },
+    "test: should not show loading state when empty completer array is provided": function(done) {
+        var editor = initEditor("");
+        editor.completers = [];
+        var completer = Autocomplete.for(editor);
+        completer.showLoadingState = true;
+
+        user.type("Ctrl-Space");
+
+        // Tooltip should not be open
+        assert.ok(!(completer.popup && completer.popup.isOpen));
+
+        done();
     }
 };